New API for account namespace and resources
